create continuous & categorical files
